change print.nif output
Previously, the first line in the print.nif() output included the name of the
analyte considered to be most relevant (i.e., as obtained from guess_analyte()).
This may have led to confusion regarding the other analytes included in the data
set, or incomprehensible output in cases where "ANALYTE" is not in the colums so
that guess_analyte() would fall back to the compartment numbers in lieu of the
analyte names.

That line has now been integrated with the info line about the number of studies
(if STUDYID is available), and a further line summarizing all analyte names was
added.